loading reducer refactor
