+++ src/Loading/reducer.js
@@ -1,5 +1,6 @@
-import * as actionTypes from '../actions/actionTypes';
-import { updateObject } from '../utility';
+import { updateObject } from '../Store/utility';
+
+import * as actionTypes from './actions';
 
 const initialState = {
   isLoading: false,
