+++ src/Loading/thunks.js
@@ -1,4 +1,4 @@
-import * as actionTypes from './actionTypes';
+import * as actionTypes from './actions';
 
 export const startLoading = () => {
   return {
