+++ src/Store/reducers/rootReducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 
 import authReducer         from '../../Auth/Reducer'
-import loadingReducer      from './loadingReducer'
+import loadingReducer      from '../../Loading/reducer'
 
 export default combineReducers({
   auth:    authReducer,
